Add installer support
