+++ .gitignore
@@ -31,7 +31,9 @@
 # Cache
 __pycache__/*
 test_package/__pycache__/*
+test_package_installer/__pycache__/*
 test_package/build/*
+test_package_installer/build/*
 build/*
 *.pyc
 tmp/
