+++ .travis.yml
@@ -8,8 +8,12 @@ linux: &linux
 osx: &osx
    os: osx
    language: generic
+stages:
+  - default  # default are all non-installer jobs
+  - installer
 matrix:
    include:
+      - stage: default
       - <<: *linux
         env: CONAN_GCC_VERSIONS=4.9 CONAN_DOCKER_IMAGE=conanio/gcc49
       - <<: *linux
@@ -46,6 +50,17 @@ matrix:
       - <<: *osx
         osx_image: xcode10.2
         env: CONAN_APPLE_CLANG_VERSIONS=10.0
+
+      - stage: installer
+        <<: *linux
+        env: CONAN_GCC_VERSIONS=7 CONAN_DOCKER_IMAGE=conanio/gcc7-centos6-x86 CONAN_ARCHS=x86 ARCH=x86 CPT_TEST_FOLDER=test_package_installer CONAN_CONANFILE=conanfile_installer.py
+      - stage: installer
+        <<: *linux
+        env: CONAN_GCC_VERSIONS=7 CONAN_DOCKER_IMAGE=conanio/gcc7-centos6 CONAN_ARCHS=x86_64 ARCH=x86_64 CPT_TEST_FOLDER=test_package_installer CONAN_CONANFILE=conanfile_installer.py
+      - stage: installer
+        <<: *osx
+        osx_image: xcode9.4
+        env: CONAN_APPLE_CLANG_VERSIONS=9.1 CONAN_ARCHS=x86_64 ARCH=x86_64 CPT_TEST_FOLDER=test_package_installer CONAN_CONANFILE=conanfile_installer.py
 
 install:
   - chmod +x .ci/install.sh
