+++ appveyor.yml
@@ -4,6 +4,17 @@ environment:
     PYTHON_HOME: "C:\\Python37"
 
     matrix:
+        - MINGW_CONFIGURATIONS: "8@x86@dwarf2@posix"
+          ARCH: x86
+          CONAN_ARCHS: x86
+          CPT_TEST_FOLDER: test_package_installer
+          CONAN_CONANFILE: conanfile_installer.py
+        - MINGW_CONFIGURATIONS: "8@x86_64@seh@posix"
+          ARCH: x86_64
+          CONAN_ARCHS: x86_64
+          CPT_TEST_FOLDER: test_package_installer
+          CONAN_CONANFILE: conanfile_installer.py
+
         - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2015
           CONAN_VISUAL_VERSIONS: 12
           CONAN_EXCLUDE_VCVARS_PRECOMMAND: 1
